move dashboard code to component-dashboard package
