+++ app/routes.js
@@ -1,6 +1,9 @@
 import React from 'react'
 import { Route, Switch } from 'react-router-dom'
 
+//Components
+import DashboardPage from '@elifesciences/component-dashboard/client'
+
 import { AuthenticatedComponent, Layout } from './components/global'
 import ErrorBoundary from './components/global/ErrorBoundary'
 import CookieNotice from './components/global/CookieNotice'
@@ -8,7 +11,6 @@ import CookieNotice from './components/global/CookieNotice'
 import LandingRedirect from './components/pages/LandingRedirect'
 import LoginPage from './components/pages/Login'
 import LogoutPage from './components/pages/Logout/index'
-import DashboardPage from './components/pages/Dashboard'
 import AuthorGuide from './components/pages/StaticPages/AuthorGuide'
 import ReviewerGuide from './components/pages/StaticPages/ReviewerGuide'
 import ContactUs from './components/pages/StaticPages/ContactUs'
