+++ packages/component-dashboard/client/DashboardContent.js
