+++ packages/component-dashboard/client/DashboardContent.test.js
