+++ packages/component-dashboard/client/ManuscriptsQuery.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Query } from 'react-apollo'
-import Loading from '../../ui/atoms/Loading'
+import Loading from 'ui/atoms//Loading'
 
 const ManuscriptsQuery = ({ query, children }) => (
   <Query fetchPolicy="cache-and-network" query={query}>
