+++ packages/component-dashboard/client/components/DashboardList.js
@@ -3,8 +3,8 @@ import styled from 'styled-components'
 import { Box } from '@rebass/grid'
 import { th } from '@pubsweet/ui-toolkit'
 
-import DashboardListItem from 'ui/molecules/DashboardListItem'
-import Paragraph from '../atoms/Paragraph'
+import DashboardListItem from './DashboardListItem'
+import Paragraph from 'ui/atoms/Paragraph'
 
 const EmptyListMessage = styled(Box)`
   text-align: center;
