+++ packages/component-dashboard/client/components/ManuscriptStatus.js
