+++ packages/component-dashboard/client/graphql/mutations.js
@@ -1,0 +1,18 @@
+import gql from 'graphql-tag'
+//TODO: component-dashboard needs its own model or a shared gql fragment for manuscript
+import { manuscriptFragment } from '../../../../app/components/pages/SubmissionWizard/operations'
+
+export const CREATE_MANUSCRIPT = gql`
+  mutation CreateManuscript {
+    createManuscript {
+      ...WholeManuscript
+    }
+  }
+  ${manuscriptFragment}
+`
+
+export const DELETE_MANUSCRIPT = gql`
+  mutation DeleteManuscript($id: ID!) {
+    deleteManuscript(id: $id)
+  }
+`
