+++ packages/component-dashboard/client/index.js
