+++ packages/component-dashboard/client/panels/Archived/index.js
