+++ packages/component-dashboard/client/panels/Submissions/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import DashboardList from 'ui/molecules/DashboardList'
-import { ALL_MANUSCRIPTS } from '../../operations'
+import DashboardList from '../../components/DashboardList'
+import { ALL_MANUSCRIPTS } from '../../graphql/queries'
 import ManuscriptsQuery from '../../ManuscriptsQuery'
 
 const Submissions = () => (
