+++ packages/component-dashboard/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@elifesciences/component-dashboard",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT"
+}
